SideNavBar Changers
Sorry, I delited yarn lock file here.
check SideNav.css I added some code there if item is in focus, means when user is going to press it and we also can add some effect for text if needed for interaction to appear on hover if you want. But I guess it's to much